refactor: 레이아웃 수정 및 라우터 정리
