--- src/components/Layout/MinimalLayout.tsx
@@ -1,9 +1,0 @@
-import type { PropsWithChildren } from 'react';
-
-import { minimalLayout } from './layout.css';
-
-const MinimalLayout = ({ children }: PropsWithChildren) => {
-  return <main className={minimalLayout}>{children}</main>;
-};
-
-export default MinimalLayout;
