+++ src/components/Layout/layout.css.ts
@@ -1,9 +1,5 @@
 import { style } from '@vanilla-extract/css';
 
-export const defaultLayout = style({
+export const main = style({
   padding: '50px 0 70px',
-});
-
-export const minimalLayout = style({
-  paddingTop: 50,
 });
